\name{lsos}
\alias{lsos}
\title{lsos}
\description{# shorthand for \code{\link{.ls.objects}}}
\usage{lsos(..., n = 10)}
\arguments{
  \item{\dots}{
}
  \item{n}{
}
}



\author{Who wrote it}




\examples{
lsos()
## Type   Size PrettySize Rows Columns
## pca.res                 PCA 790128   771.6 Kb    7      NA
## DF               data.frame 271040   264.7 Kb  669      50
}
